requirements.txt
fire==0.6.0
numpy==1.26.4
PyYAML==6.0.1
scipy==1.13.1
soundfile==0.12.1
tqdm==4.66.4
